+++ index.js
@@ -48,7 +48,7 @@ export { default as YAxisControlsSettings } from "./components/playback/controls
 export { default as DesktopPlayerControls } from "./components/playback/controls/player/DesktopPlayerControls.vue";
 export { default as PlayerControlsSettingsButton } from "./components/playback/controls/player/PlayerControlsSettingsButton.vue";
 export { default as PlateBarcode } from "./components/playback/controls/PlateBarcode.vue";
-export { default as AdditionalControls } from "./components/playback/controls/AdditionalControls.vue";
+export { default as StimulationControls } from "./components/playback/controls/StimulationControls.vue";
 
 // Settings Components
 export { default as SettingsForm } from "./components/settings/SettingsForm.vue";