+++ tests/e2e/playback/controls/additional-controls.e2e.js
@@ -14,14 +14,14 @@ fixture`playback/controls/additional-controls/basic`
 
 const status_btn_active = Selector(".span__stimulation-controls-play-stop-button--active");
 
-test("testing the AdditionalControls layout on initialization", async (t) => {
+test("testing the StimulationControls layout on initialization", async (t) => {
   const screenshot_path_base = path.join("playback", "controls", "additional-controls");
   const screenshot_path = path.join(screenshot_path_base, "basic-init-disabled");
 
   await testcafe_page_visual_regression(t, screenshot_path);
 });
 
-test("testing the AdditionalControls layout after enabled", async (t) => {
+test("testing the StimulationControls layout after enabled", async (t) => {
   const screenshot_path_base = path.join("playback", "controls", "additional-controls");
   const screenshot_path = path.join(screenshot_path_base, "basic-init-enabled");
   await t.click(enable_button);
@@ -29,7 +29,7 @@ test("testing the AdditionalControls layout after enabled", async (t) => {
   await testcafe_page_visual_regression(t, screenshot_path);
 });
 
-test("testing the AdditionalControls layout when stimulation is active", async (t) => {
+test("testing the StimulationControls layout when stimulation is active", async (t) => {
   const screenshot_path_base = path.join("playback", "controls", "additional-controls");
   const screenshot_path = path.join(screenshot_path_base, "basic-active");
   await t.click(enable_button).click(update_button).click(status_btn_active).wait(1000);