+++ tests/unit/components/playback/controls/AdditionalControls.spec.js
@@ -1,6 +1,6 @@
 import Vuex from "vuex";
 import { createLocalVue, mount } from "@vue/test-utils";
-import AdditionalControls from "@/components/playback/controls/AdditionalControls.vue";
+import StimulationControls from "@/components/playback/controls/StimulationControls.vue";
 import playback_module from "@/store/modules/playback";
 
 describe("store/stimulation", () => {
@@ -22,9 +22,9 @@ describe("store/stimulation", () => {
     jest.clearAllMocks();
   });
 
-  describe("AdditionalControls", () => {
-    test("When AdditionalControls mounts, Then the initial play state should be false", () => {
-      const wrapper = mount(AdditionalControls, {
+  describe("StimulationControls", () => {
+    test("When StimulationControls mounts, Then the initial play state should be false", () => {
+      const wrapper = mount(StimulationControls, {
         store,
         localVue,
       });
@@ -38,7 +38,7 @@ describe("store/stimulation", () => {
       });
 
       test("Then controls block should be displayed", () => {
-        const wrapper = mount(AdditionalControls, {
+        const wrapper = mount(StimulationControls, {
           store,
           localVue,
         });
@@ -53,7 +53,7 @@ describe("store/stimulation", () => {
       });
 
       test("Then controls block should not be displayed", () => {
-        const wrapper = mount(AdditionalControls, {
+        const wrapper = mount(StimulationControls, {
           store,
           localVue,
         });
@@ -63,21 +63,23 @@ describe("store/stimulation", () => {
 
       test("Given a stimulation is active, When a user clicks the button to turn off stimulation, Then a signal should be dispatched to BE", async () => {
         const dispatch_spy = jest.spyOn(store, "dispatch");
-        dispatch_spy.mockImplementation(async () => await store.commit("stimulation/set_stim_status", false));
+        dispatch_spy.mockImplementation(
+          async () => await store.commit("stimulation/set_stim_play_state", false)
+        );
 
         store.state.stimulation.protocol_assignments = { test: "assignment" };
-        const wrapper = mount(AdditionalControls, {
+        const wrapper = mount(StimulationControls, {
           store,
           localVue,
         });
 
         wrapper.vm.play_state = true;
         await wrapper.find(".span__stimulation-controls-play-stop-button--active").trigger("click");
-        expect(dispatch_spy).toHaveBeenCalledWith("stimulation/stop_stim_status");
+        expect(dispatch_spy).toHaveBeenCalledWith("stimulation/stop_stimulation");
       });
 
       test("Given there are no wells assigned with a protocol, When a user clicks to start a stimulation, Then no signal should be dispatched to BE", async () => {
-        const wrapper = mount(AdditionalControls, {
+        const wrapper = mount(StimulationControls, {
           store,
           localVue,
         });
@@ -89,10 +91,12 @@ describe("store/stimulation", () => {
 
       test("Given a stimulation is inactive and there are protocol assigned wells, When a user clicks the button to turn on stimulation, Then a signal should be dispatched to BE", async () => {
         const dispatch_spy = jest.spyOn(store, "dispatch");
-        dispatch_spy.mockImplementation(async () => await store.commit("stimulation/set_stim_status", true));
+        dispatch_spy.mockImplementation(
+          async () => await store.commit("stimulation/set_stim_play_state", true)
+        );
 
         store.state.stimulation.protocol_assignments = { test: "assignment" };
-        const wrapper = mount(AdditionalControls, {
+        const wrapper = mount(StimulationControls, {
           store,
           localVue,
         });
@@ -101,16 +105,16 @@ describe("store/stimulation", () => {
         expect(wrapper.vm.play_state).toBe(true);
         expect(dispatch_spy).toHaveBeenCalledWith("stimulation/create_protocol_message");
       });
-      test("When set_stim_status is called with different values, Then current gradient is updated correctly", async () => {
-        const wrapper = mount(AdditionalControls, {
+      test("When set_stim_play_state is called with different values, Then current gradient is updated correctly", async () => {
+        const wrapper = mount(StimulationControls, {
           store,
           localVue,
         });
 
-        store.commit("stimulation/set_stim_status", true);
+        store.commit("stimulation/set_stim_play_state", true);
         await wrapper.vm.$nextTick(); // wait for update
         expect(wrapper.vm.current_gradient).toStrictEqual(wrapper.vm.active_gradient);
-        store.commit("stimulation/set_stim_status", false);
+        store.commit("stimulation/set_stim_play_state", false);
         await wrapper.vm.$nextTick(); // wait for update
         expect(wrapper.vm.current_gradient).toStrictEqual(wrapper.vm.inactive_gradient);
       });